Move and fix Live OS init source test
Create a new module_payload_sources_test test source file to test
sources. If it starts to grow too much it can be split.
