+++ tests/nosetests/pyanaconda_tests/module_payload_live_os_test.py
@@ -24,10 +24,7 @@ from mock import Mock, patch
 from tests.nosetests.pyanaconda_tests import check_task_creation, patch_dbus_publish_object
 
 from pyanaconda.core.constants import INSTALL_TREE
-from pyanaconda.dbus.typing import get_native
 from pyanaconda.modules.common.constants.objects import LIVE_OS_HANDLER
-from pyanaconda.modules.common.errors.payload import SourceSetupError
-from pyanaconda.modules.common.structures.storage import DeviceData
 from pyanaconda.modules.common.task.task_interface import TaskInterface
 from pyanaconda.modules.payload.base.initialization import PrepareSystemForInstallationTask, \
     CopyDriverDisksFilesTask
@@ -183,79 +180,3 @@ class LiveOSHandlerInterfaceTestCase(unittest.TestCase):
             self.assertEqual(object_path, task_paths[i])
             self.assertIsInstance(obj, TaskInterface)
             self.assertIsInstance(obj.implementation, task_classes[i])
-
-
-class LiveOSHandlerTasksTestCase(unittest.TestCase):
-
-    def setUp(self):
-        self.live_os_module = LiveOSHandlerModule()
-        self.live_os_interface = LiveOSHandlerInterface(self.live_os_module)
-
-        self.callback = Mock()
-        self.live_os_interface.PropertiesChanged.connect(self.callback)
-
-    @patch("pyanaconda.modules.payload.live.initialization.mount")
-    @patch("pyanaconda.modules.payload.live.initialization.stat")
-    @patch("os.stat")
-    @patch("pyanaconda.dbus.DBus.get_proxy")
-    def setup_install_source_task_test(self, proxy_getter, os_stat, stat, mount):
-        """Test Live OS setup installation source task."""
-        device_tree = Mock()
-        proxy_getter.return_value = device_tree
-        device_tree.ResolveDevice = Mock()
-        device_tree.ResolveDevice.return_value = "resolvedDeviceName"
-
-        device = DeviceData()
-        device.path = "/resolved/path/to/base/image"
-
-        device_tree.GetDeviceData = Mock()
-        device_tree.GetDeviceData.return_value = get_native(DeviceData.to_structure(device))
-
-        mount.return_value = 0
-
-        SetUpInstallationSourceTask(
-            "/path/to/base/image",
-            "/path/to/mount/source/image"
-        ).run()
-
-        device_tree.ResolveDevice.assert_called_once_with("/path/to/base/image")
-        os_stat.assert_called_once_with("/resolved/path/to/base/image")
-
-    @patch("pyanaconda.dbus.DBus.get_proxy")
-    def setup_install_source_task_missing_image_test(self, proxy_getter):
-        """Test Live OS setup installation source task missing image error."""
-        device_tree = Mock()
-        proxy_getter.return_value = device_tree
-        device_tree.ResolveDevice = Mock()
-        device_tree.ResolveDevice.return_value = ""
-
-        with self.assertRaises(SourceSetupError):
-            SetUpInstallationSourceTask(
-                "/path/to/base/image",
-                "/path/to/mount/source/image"
-            ).run()
-
-    @patch("pyanaconda.modules.payload.live.initialization.mount")
-    @patch("pyanaconda.modules.payload.live.initialization.stat")
-    @patch("os.stat")
-    @patch("pyanaconda.dbus.DBus.get_proxy")
-    def setup_install_source_task_failed_to_mount_test(self, proxy_getter, os_stat, stat, mount):
-        """Test Live OS setup installation source task mount error."""
-        device_tree = Mock()
-        proxy_getter.return_value = device_tree
-        device_tree.ResolveDevice = Mock()
-        device_tree.ResolveDevice.return_value = "resolvedDeviceName"
-
-        device = DeviceData()
-        device.path = "/resolved/path/to/base/image"
-
-        device_tree.GetDeviceData = Mock()
-        device_tree.GetDeviceData.return_value = get_native(DeviceData.to_structure(device))
-
-        mount.return_value = -20
-
-        with self.assertRaises(SourceSetupError):
-            SetUpInstallationSourceTask(
-                "/path/to/base/image",
-                "/path/to/mount/source/image"
-            ).run()
